% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{av}
\alias{av}
\title{Atomize}
\usage{
av(...)
}
\arguments{
\item{...}{Objects to atomized.}
}
\value{
An atomic vector
}
\description{
Flattens \code{...} to an atomic vector of atomic constituent values.
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{choose_from}()},
\code{\link{f0}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{fs_t}()},
\code{\link{g0}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{is_err}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{p0}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
