% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{callers}
\alias{callers}
\alias{caller}
\title{Function names and counts in the call stack}
\usage{
callers(.vec = TRUE)

caller(.scl = TRUE)
}
\arguments{
\item{.vec}{\code{TRUE} or \code{FALSE} indicating whether to represent both package and function in a character vector rather than as a two element list with one element for packages and another for functions.}

\item{.scl}{\code{TRUE} or \code{FALSE} indicating whether to collapse package and function into a character scalar rather than as a two element list with one element for packages and another for functions.}

\item{n}{A \link[cmp_psw_vec]{complete positive whole-number scalar} giving the number of generations back in the function call stack to go.}

\item{.err}{\code{TRUE} or \code{FALSE} indicating whether to throw an error if one is encountered.}

\item{...}{An arbitrary number of \link[=cmp_psw_vec]{complete positive whole-number vecs} giving the number(s) of generations back in the function call stack to go.}
}
\value{
\strong{A \verb{list(pkg = <character vector>, fun = <character vector>)}} \cr\cr  When \verb{vec = FALSE} or \verb{scl = FALSE}.
\cr\cr  \strong{A character vector} (when \code{vec = TRUE})                       \cr\cr \code{callers}
\cr\cr  \strong{A character scalar} (when \code{scl = TRUE})                       \cr\cr \code{caller}
}
\description{
This family of functions addresses the function call stack \emph{from the perspective of a function that calls one of the functions in this family}. Consistent with that perspective, the basic terminology this family of functions uses is the following:
\tabular{ll}{  \strong{Terminology}   \tab \strong{Function identity} \cr
  "command line"    \tab The command line      \cr
  "caller"          \tab Parent function       \cr
  "callers"         \tab All parent functions    }
}
\examples{
egD <- function() {list(caller    = caller()                 ,
                        callers   = callers()                )}
egC <- function() {list(caller    = caller()                 ,
                        callers   = callers()                ,
                        egD       = egD()                    )}
egB <- function() {list(caller    = caller()                 ,
                        callers   = callers()                ,
                        egC       = egC()                    )}
egA <- function() {list(caller    = caller()                 ,
                        callers   = callers()                ,
                        egB       = egB()                    )}
egA()
av(egA())
}
\seealso{
Other meta: 
\code{\link{console_basics}()}
}
\concept{environments}
\concept{meta}
