% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog.R
\encoding{UTF-8}
\name{dialog}
\alias{dialog}
\alias{alert}
\alias{acknowledge}
\alias{choose1}
\alias{chooseN}
\alias{NO}
\alias{YES}
\alias{OK}
\alias{CANCEL}
\alias{ask}
\alias{ask_new}
\alias{choose_dir}
\alias{choose_doc}
\title{Console-Based User Dialog (may also use package \code{svDialogs} functions)}
\usage{
alert(
  ...,
  .title = "alert",
  .sub = "",
  .ps = "",
  .def = "",
  .ft = "r|w|b",
  .fs = "k|y|p",
  .fm = "",
  .fp = "k|y|i",
  .d = " ",
  .clear = FALSE
)

acknowledge(
  ...,
  .sub = "",
  .ps = "",
  .ft = "r|w|b",
  .fs = "k|y|p",
  .fp = "k|y|i",
  .d = " ",
  .clear = FALSE
)

choose1(options, ..., .ft = "r|w|b", .fs = "k|y|p", .d = " ", .clear = FALSE)

chooseN(
  options,
  ...,
  .all = TRUE,
  .none = FALSE,
  .n = NULL,
  .min = NULL,
  .max = NULL,
  .ft = "r|w|b",
  .fs = "k|y|p",
  .d = " ",
  .clear = FALSE
)

NO(..., .ft = "r|w|b", .fs = "k|y|p", .d = " ", .clear = FALSE)

YES(..., .ft = "r|w|b", .fs = "k|y|p", .d = " ", .clear = FALSE)

OK(..., .ft = "r|w|b", .fs = "k|y|p", .d = " ", .clear = FALSE)

CANCEL(..., .ft = "r|w|b", .fs = "k|y|p", .d = " ", .clear = FALSE)

ask(
  ...,
  Default = "",
  .sub = "",
  .ft = "r|w|b",
  .fs = "k|y|p",
  .fm = "",
  .fp = "k|y|i",
  .d = " ",
  .clear = FALSE
)

ask_new(
  old,
  type = "replacement values",
  .u = TRUE,
  .sub = "",
  .ft = "r|w|b",
  .fs = "k|y|p",
  .fm = "",
  .fp = "k|y|i",
  .d = "|",
  .clear = FALSE
)

choose_dir(
  dir.type = "directory",
  .sub = "",
  .ft = "r|w|b",
  .fs = "k|y|p",
  .fm = "",
  .fp = "k|y|i",
  .d = " ",
  .clear = FALSE
)

choose_doc(
  doc.type = "document",
  .sub = "",
  .ft = "r|w|b",
  .fs = "",
  .fm = "",
  .fp = "k|y|i",
  .d = " ",
  .clear = FALSE
)
}
\arguments{
\item{...}{An arbitrary, optional number of arguments which are \link[=av]{atomized} into a character scalar message to be posted to the console.}

\item{.title}{A .title for the alert (\code{type = 'error'} results in stopping execution after posting the alert). Should be short to avoid formatting problems.}

\item{.sub}{A character scalar alert subtitle to post to the console on the line following the title Should be short to avoid formatting problems.}

\item{.ps}{A character scalar suffix to post to the console on the line following the alert message contained in \code{...}.}

\item{.def}{A character scalar containing a default message if \link[=av]{atomizing} and collapsing \code{...} to a character scalar results in a blank string (\code{""}).}

\item{.ft, .fs, .fm, .fp}{Formatting values consistent with the description in the \emph{specifying formats} section giving formatting instructions for, respectively, the title (in \code{.title}), subtitle (in \code{.sub}), message (in \code{...}), and postscript (in \code{.ps}).}

\item{.d}{A character scalar delimiter for collapsing \code{...} args into a character scalar.}

\item{.clear}{A non-\code{NA} logical scalar indicating whether to clear the console before each interaction with the user.}

\item{options}{An atomic vector listing options to choose from.}

\item{.all}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add an \code{{ ALL }} value to \code{options}.}

\item{.none}{Scalar \code{TRUE} or \code{FALSE} indicating whether to add a \code{{ NONE }} value to \code{options} (implying that it is valid to select none).}

\item{.n}{An optional \link[ppp:cmp_psw_scl]{complete positive numeric whole-number scalar} (?cmp_psw_scl) indicating the number of options that must be selected. Must be contained in \code{1:length(options)}.}

\item{.min}{An optional \link[ppp:CMP]{complete} \link[ppp:PSW]{positive numeric whole-number} \link[ppp:SCL]{scalar} indicating the minimum number of options that may be selected. Must be \code{NULL} when \code{.n} is non-\code{NULL}.}

\item{.max}{An optional complete positive numeric whole-number scalar indicating the maximum number of options that may be selected. Must be \code{NULL} when \code{.n} is non-\code{NULL}.}

\item{old}{A \link[=chr_vec]{character vec} of unique values to be replaced.}

\item{type}{A character scalar describing the type of replacement values to be entered.}

\item{.u}{Scalar \code{TRUE} or \code{FALSE} indicating whether replacement values must be unique.}
}
\value{
\strong{The} \code{NULL} \strong{object} \cr\cr \verb{acknowledge, alert}
\cr\cr  \strong{A character scalar}    \cr\cr \code{choose_dir} (a directory path) \cr \code{choose_doc} (a document path) \cr \code{ask}
\cr\cr  \strong{A character vector}    \cr\cr \code{ask_new}
\cr\cr  \strong{An atomic vector}      \cr\cr \code{chooseN}
\cr\cr  \strong{An atomic scalar}      \cr\cr \code{choose1}
\cr\cr  \strong{A logical scalar}      \cr\cr \verb{CANCEL, YES, NO, OK}
}
\description{
All functions \link[=collapse_dots]{collapses} \code{...} args into a prompt. Each posts an alert to the console, posts the prompt (if any), followed by a specific action.
}
\details{
\tabular{ll}{  \code{acknowledge}   \tab Waits for user to acknowledge.                        \cr
                  \tab                                                       \cr
  \verb{OK, CANCEL}    \tab Offers OK and CANCEL options.\eqn{^{(3)}}             \cr
                  \tab                                                       \cr
  \code{choose_doc}    \tab Asks user to choose a document.\eqn{^{(1)}}           \cr
  \code{choose_dir}    \tab Asks user to choose a directory.\eqn{^{(1)}}          \cr
                  \tab                                                       \cr
  \code{choose1}       \tab Asks user to choose \code{1} option.\eqn{^{(1)}}           \cr
  \code{chooseN}       \tab Asks user to choose \verb{1+} options.\eqn{^{(1,2)}}       \cr
                  \tab                                                       \cr
  \verb{YES, NO}       \tab Offers YES, NO, and CANCEL options.\eqn{^{(1,3)}}     \cr
                  \tab                                                       \cr
  \code{ask_new}       \tab Asks for a list of new values.\eqn{^{(1)}}            \cr
                  \tab                                                       \cr
  \code{alert}         \tab No subsequent action.                                 \cr
                  \tab                                                       \cr
  \code{ask}           \tab Asks for typed input.\eqn{^{(1)}}                     \cr
                  \tab                                                         }
\tabular{l}{  \eqn{^{(1)}} Stops execution if user chooses \code{CANCEL}.                     \cr
  \eqn{^{(2)}} Stops execution if \code{stop | (type = "error" & is.null(stop))}. \cr
  \eqn{^{(3)}} Returns \code{TRUE} if user choice matches the function name.        }
The following give templates for what the user sees, where any value derived from arguments will be absent if it is or resolves to a blank string.
\cr\cr
\strong{\code{alert}}
\tabular{ll}{
  \strong{CONDITION}           \tab \strong{VALUE}                       \cr
  \code{g(.d, .title) != ""}   \tab formatted title from \code{.title}.  \cr
  \code{g(.d, .sub) != ""}     \tab formatted subtitle from \code{.sub}. \cr
  \code{g(.d, ...) != ""}      \tab formatted message from \code{...}.   \cr
                          \tab \verb{< blank line >}                \cr
  \code{g(.d, .ps) != ""}      \tab formatted postscript from \code{.ps}.  }
.all other templates incorporate a call to \code{alert(.)} with components as shown below.
\cr\cr\strong{\code{acknowledge}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                                 \cr
  .title          \tab \code{'ACKNOWLEDGMENT REQUIRED'}               \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}                       \cr
  Message         \tab \code{'< g(.d, ...) >'}                        \cr
                  \tab \verb{< blank line >}                          \cr
  Postscript      \tab \code{'press [return] or [enter] to continue:'}  }
\strong{\code{choose1}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                         \cr
  .title          \tab \code{'RESPONSE REQUIRED'}             \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}               \cr
  Message         \tab \code{'< g(.d, ...) >'}                \cr
                  \tab \verb{< blank line >}                  \cr
  Postscript      \tab \code{'choose an option - Select one'} \cr
                  \tab \verb{< blank line >}                  \cr
  Cancel.Option   \tab \verb{'1: \{ CANCEL \}}'                 \cr
  Option.1        \tab \code{'2: < options[1] >'}             \cr
  Option.2        \tab \code{'3: < options[2] >'}             \cr
  ...             \tab ...                               \cr
  Option.n        \tab \code{'< n + 1 >: < options[n] >'}     \cr
  Prompt          \tab \code{'Selection:'}                      }
\strong{\code{chooseN}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                                                                      \cr
  .title           \tab \code{'RESPONSE REQUIRED'}                                                         \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}                                                            \cr
  Message         \tab \code{'< g(.d, ...) >'}                                                             \cr
                  \tab \verb{< blank line >}                                                               \cr
  Postscript      \tab \code{'choose < n | between n1 and n2 > options - Select one or more'}              \cr
                  \tab \verb{< blank line >}                                                               \cr
  Cancel.Option   \tab \code{'1: { CANCEL }'}                                                              \cr
  Option.1        \tab \code{'2: < options[1] >'}                                                          \cr
  Option.2        \tab \code{'3: < options[2] >'}                                                          \cr
  ...             \tab ...                                                                            \cr
  Option.n        \tab \code{'< n + 1 >: < options[n] >'}                                                  \cr
  ALL.option      \tab \code{'< n + 2 >: { ALL }'}                                                         \cr
  NONE.option     \tab \code{'< n + 3 >: { NONE }'}                                                        \cr
  Prompt          \tab \code{'Enter one or more numbers separated by spaces and then ENTER, or 0 to cancel'} }
\strong{\code{NO}} and \strong{\code{YES}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                         \cr
  Title           \tab \code{'RESPONSE REQUIRED'}             \cr
  Subtitle        \tab '\verb{< g(.d, .sub) >'}               \cr
  Message         \tab \code{'< g(.d, ...) >'}                \cr
                  \tab \verb{< blank line >}                  \cr
  Postscript      \tab \code{'choose an option - Select one'} \cr
                  \tab \verb{< blank line >}                  \cr
  Cancel.option   \tab \code{'1: { CANCEL }'}                 \cr
  Yes.option      \tab \code{'2: { YES }'}                    \cr
  No.option       \tab \code{'3: { NO }'}                     \cr
  Prompt          \tab \code{'Selection:'}                      }
\strong{\code{OK}} and \strong{\code{CANCEL}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                         \cr
  Title           \tab \code{'RESPONSE REQUIRED'}             \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}               \cr
  Message         \tab \code{'< g(.d, ...) >'}                \cr
                  \tab \verb{< blank line >}                  \cr
  Postscript      \tab \code{'choose an option - Select one'} \cr
                  \tab \verb{< blank line >}                  \cr
  Cancel.option   \tab \code{'1: { CANCEL }'}                 \cr
  OK.option       \tab \code{'2: { OK }'}                     \cr
  Prompt          \tab \code{'Selection: '}                     }
\strong{\code{ask}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                 \cr
  Title           \tab \code{'RESPONSE REQUIRED'}     \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}       \cr
  Message         \tab \code{'< g(.d, ...) >'}        \cr
                  \tab \verb{< blank line >}          \cr
  Postscript      \tab \code{'enter your response: '}   }
\strong{\code{ask_new}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                                                                                                                \cr
  .title           \tab \code{'RESPONSE REQUIRED'}                                                                                                   \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}                                                                                                      \cr
  Message         \tab \code{'Enter a pipe-separated list of < N(options) > replacement values for the following pipe-separated original values: '}  \cr
                  \tab \verb{< blank line >}                                                                                                         \cr
  old.values      \tab \code{'< paste0(old, collapse = " | ") >'}                                                                                    \cr
                  \tab \verb{< blank line >}                                                                                                         \cr
  Postscript      \tab \code{'enter your response: '}                                                                                                  }
\strong{\code{choose_dir}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                                          \cr
  Title           \tab \code{'ACKNOWLEDGMENT REQUIRED'}                        \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}                                \cr
  Message         \tab \code{'In the next dialog box, select a < dir.type >.'}   }
\strong{\code{choose_doc}}
\tabular{ll}{
  \strong{COMPONENT}   \tab \strong{VALUE}                                          \cr
  Title           \tab \code{'ACKNOWLEDGMENT REQUIRED'}                        \cr
  Subtitle        \tab \code{'< g(.d, .sub) >'}                                \cr
  Message         \tab \code{'In the next dialog box, select a < doc.type >.'}   }
}
\section{Specifying formats}{
 When formatting arguments (\code{.ft}, \code{.fs}, \code{.fm}, and \code{.fp}) take the special value \code{""}, the corresponding alert elements (\code{.title}, \code{.sub}, \code{...}, and \code{.ps}, respectively) are posted to the console without special formatting.
\cr\cr Otherwise, formatting arguments must be \link[=cmp_str_vec]{complete string vecs} that when \link[=av]{atomized} and \link[=ssplit]{split along pipes} results in a three-element character vector, the first element of which is used to specify \link[=bg]{text background color} and must be a value from \code{\link{bg_vals}()}, the second element of which is used to specify \link[=fg]{text foreground color} and must be a value from \code{\link{fg_vals}()}, and the last of which specifies \link[=st]{text style} and must be a value from \code{\link{st_vals}()}.
}

\examples{
\dontrun{

  ## because formatting doesn't show up in help viewer examples output

  egA    <- "two-part"
  egB    <- "message"

  egFT   <- c("yellow", "red", "plain")
  egFS   <- c("blk|wht", "und")
  egFM   <- "b|y|i"

  egT    <- ".title"
  egS    <- "Subtitle"

  egOpts <- paste("option", letters[1:10])

  egM1   <- "Do you want to continue?"
  egM2   <- "Why do you want to continue?"

  alert(egA, egB, .d = " ")
  alert(egA, egB, .title = egTitle, .d = " ")
  alert(egA, egB, .sub = l, .d = " ")
  alert(egA, egB, .title = egTitle, .sub = egSub, .d = " ")
  alert(egA, egB, .title = egTitle, .sub = egSub, .d = " ", .fm = egFM)
  alert(egA, egB, .title = egTitle, .sub = egSub, .d = " ", .fs = egFS)
  alert(egA, egB, .title = egTitle, .sub = egSub, .d = " ", .ft = egFT)
  alert(egA, egB, .title = egTitle, .sub = egSub, .d = " ", .ft = egFT, .fs = egFS, .fm = egFM)
  alert(.title = egT, .sub = egSub, .ft = egFT, .fs = egFS)
  alert(.title = egT, .sub = egS)
  alert(.sub = egS, .fs = egFS)

  acknowledge(egA, egB)

  choose1(egOpts)
  chooseN(egOpts)
  chooseN(egOpts, .all = T, .none = F, Min = 6, Max = 10)

  NO(egMsg1)
  OK(egMsg2)
  YES(egMsg1)
  CANCEL(egMsg1)

  ask(Msg2)

  ask_new(egOpts)

  choose_dir(dir.type = "directory for R scripts")

  choose_doc(doc.type = "document to read")

}
}
\seealso{
Other user: 
\code{\link{console_basics}}
}
\concept{dialogs}
\concept{user}
