% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.fmt_errs}
\alias{.fmt_errs}
\title{Validate Formatting Argument Values}
\usage{
.fmt_errs(
  .st = NULL,
  .bg = NULL,
  .fg = NULL,
  .d = " ",
  .null.st = TRUE,
  .null.bg = TRUE,
  .null.fg = TRUE
)
}
\arguments{
\item{.st}{Optional style spec.}

\item{.bg}{Optional background color spec.}

\item{.fg}{Optional foreground color spec.}

\item{.d}{Character scalar text delimiter.}

\item{.null.st}{Logical scalar indicating whether a \code{NULL} style spec is valid.}

\item{.null.bg}{Logical scalar indicating whether a \code{NULL} background color spec is valid.}

\item{.null.fg}{Logical scalar indicating whether a \code{NULL} foreground color spec is valid.}
}
\description{
Validate Formatting Argument Values
}
\examples{
.fmt_errs()
.fmt_errs(.d = 100)
.fmt_errs(.st = "", .bg = "", .fg = "")
.fmt_errs(.st = "q", .bg = "q", .fg = "q")
.fmt_errs(.null.st = F, .null.bg = F, .null.fg = F)
}
\seealso{
Other utils: 
\code{\link{.cat_choose_list}()},
\code{\link{.glue_args}()},
\code{\link{.glu}()},
\code{\link{.ok_fmt}()},
\code{\link{.ssplit}()},
\code{\link{.trm}()},
\code{\link{match_alias}()}
}
\concept{utils}
