% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.glu}
\alias{.glu}
\title{Collapse and Glue \code{...} Arguments with Space as Default Delimiter}
\usage{
.glu(..., .d = " ")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.glue_args]{.glue_args}}
  \describe{
    \item{\code{}}{}
  }}

\item{.d}{Atomic scalar delimiter.}
}
\value{
A character scalar.
}
\description{
Collapse and Glue \code{...} Arguments with Space as Default Delimiter
}
\examples{
.glu(letters)
.glu(letters, 0:9, .d = "")
}
\seealso{
Other utils: 
\code{\link{.cat_choose_list}()},
\code{\link{.fmt_errs}()},
\code{\link{.glue_args}()},
\code{\link{.ok_fmt}()},
\code{\link{.ssplit}()},
\code{\link{.trm}()},
\code{\link{match_alias}()}
}
\concept{utils}
