% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.glue_args}
\alias{.glue_args}
\title{Flattens \code{...} Args to a Character Vector and Paste the Resulting Elements together Using Delimiter \code{.d}}
\usage{
.glue_args(.d, ...)
}
\arguments{
\item{.d}{Atomic scalar delimiter.}

\item{...}{Objects to be flattened.}
}
\value{
A character scalar.
}
\description{
Flattens \code{...} Args to a Character Vector and Paste the Resulting Elements together Using Delimiter \code{.d}
}
\examples{
.glue_args(" ", letters)
.glue_args("" , letters)
.glue_args("|", letters, LETTERS, 0:9)
}
\seealso{
Other utils: 
\code{\link{.cat_choose_list}()},
\code{\link{.fmt_errs}()},
\code{\link{.glu}()},
\code{\link{.ok_fmt}()},
\code{\link{.ssplit}()},
\code{\link{.trm}()},
\code{\link{match_alias}()}
}
\concept{utils}
