% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.ssplit}
\alias{.ssplit}
\title{Split Strings along Pipes and Trim Whitespace from Both Sides of All Resulting Elements}
\usage{
.ssplit(x)
}
\arguments{
\item{x}{An object of mode character.}
}
\value{
A character vector
}
\description{
Split Strings along Pipes and Trim Whitespace from Both Sides of All Resulting Elements
}
\examples{
.ssplit(" Just | a | little | thing ")
.ssplit("  Just  |  a  |  little  |  thing  ")
}
\seealso{
Other utils: 
\code{\link{.cat_choose_list}()},
\code{\link{.fmt_errs}()},
\code{\link{.glue_args}()},
\code{\link{.glu}()},
\code{\link{.ok_fmt}()},
\code{\link{.trm}()},
\code{\link{match_alias}()}
}
\concept{utils}
