% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{.trm}
\alias{.trm}
\title{Glue and Trim Whitespace from Both Sides}
\usage{
.trm(..., .d = " ")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=.glue_args]{.glue_args}}
  \describe{
    \item{\code{}}{}
  }}

\item{.d}{Atomic scalar delimiter.}
}
\value{
A character scalar.
}
\description{
Glue and Trim Whitespace from Both Sides
}
\examples{
.trm(" Just" ,  "a" ,  "little" ,  "thing ")
.trm(" Just ", " a ", " little ", " thing ")
.trm(" ", letters, LETTERS, 0:9, " ", .d = "")
}
\seealso{
Other utils: 
\code{\link{.cat_choose_list}()},
\code{\link{.fmt_errs}()},
\code{\link{.glue_args}()},
\code{\link{.glu}()},
\code{\link{.ok_fmt}()},
\code{\link{.ssplit}()},
\code{\link{match_alias}()}
}
\concept{utils}
