% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{f0}
\alias{f0}
\title{Enhancements of \code{\link[base]{ifelse}}.}
\usage{
f0(x, y, n)
}
\arguments{
\item{x}{A logical scalar (if not,\code{x} it is replaced by \code{FALSE}).}

\item{y, n}{Any valid R object.}

\item{.d}{A character scalar delimiter for collapsing objects into scalar character objects. If \code{.d} is not a character scalar, it is replaced by \code{" "}.}

\item{.cond}{A character scalar in \code{c('all', 'any', 'none')}. If \code{.cond} is not of an allowed value, it is replaced by \code{'all'}.}
}
\value{
An arbitrary object (either \code{y} or \code{n}).
}
\description{
If \code{x} is scalar \code{TRUE}, returns \code{y}. If \code{x} is anything else, returns \code{n}.
}
\examples{
f0(NA, data.frame(letters = letters), 0:26)
f0(TRUE, data.frame(letters = letters), 0:26)
f0(FALSE, data.frame(letters = letters), 0:26)
f0(list(1, "a"), data.frame(letters = letters), 0:26)
f0(c(.bad.var.name.), data.frame(letters = letters), 0:26)
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{av}()},
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{fs_t}()},
\code{\link{g0}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{is_err}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{p0}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
