% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{failsafe}
\alias{failsafe}
\title{Failsafe Evaluation of an Object}
\usage{
failsafe(x, .def = "dlg.failsafe.err")
}
\arguments{
\item{x}{Any object/expression to be evaluated, whether or not doing so produces an error.}

\item{.def}{A character scalar default error message if forcing evaluation produces an error. If not a character scalar, it is replaced with the default.}
}
\value{
Either \code{x} or the object \code{.def} with the attribute \code{'stack'} composed of the function call stack as a character vector.
}
\description{
Evaluate objects, test if an object has certain properties, and conduct binary logical operations safely. these functions never stop execution; they always produce a valid result, even if that result is an error object.
}
\examples{
failsafe(non.existent.variable)
failsafe(pi)
}
\concept{errs}
\concept{extensions}
\concept{logicals}
