% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{fs_t}
\alias{fs_t}
\title{Failsafe Evaluation of Whether \code{x} Is Scalar \code{TRUE}}
\usage{
fs_t(x)
}
\arguments{
\item{x}{An object to be evaluated for scalar \code{TRUE}-ness.}
}
\value{
A logical scalar
}
\description{
Failsafe Evaluation of Whether \code{x} Is Scalar \code{TRUE}
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{av}()},
\code{\link{choose_from}()},
\code{\link{f0}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{g0}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{is_err}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{p0}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
