% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{g0}
\alias{g0}
\title{\code{paste0(av(...), collapse = "")}}
\usage{
g0(...)
}
\value{
A character scalar.
}
\description{
\code{paste0(av(...), collapse = "")}
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{av}()},
\code{\link{choose_from}()},
\code{\link{f0}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{fs_t}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{is_err}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{p0}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
