% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{is_err}
\alias{is_err}
\title{Checks Whether Evaluation \code{x} Produces and Error}
\usage{
is_err(x)
}
\arguments{
\item{x}{An object to be evaluated.}
}
\value{
A logical scalar
}
\description{
Checks Whether Evaluation \code{x} Produces and Error
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{av}()},
\code{\link{choose_from}()},
\code{\link{f0}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{fs_t}()},
\code{\link{g0}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{p0}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
