% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{p0}
\alias{p0}
\title{Thin Wrapper for \code{\link[base]{paste0}}}
\usage{
p0(..., collapse = NULL, recycle0 = FALSE)
}
\value{
A positive integer scalar.
}
\description{
Thin Wrapper for \code{\link[base]{paste0}}
}
\seealso{
Other utils: 
\code{\link{N}()},
\code{\link{av}()},
\code{\link{choose_from}()},
\code{\link{f0}()},
\code{\link{fmt_errs}()},
\code{\link{fmt_stack}()},
\code{\link{fs_t}()},
\code{\link{g0}()},
\code{\link{glu}()},
\code{\link{g}()},
\code{\link{is_err}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
