% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{re_exports}
\alias{re_exports}
\alias{g}
\alias{N}
\alias{av}
\alias{f0}
\alias{g0}
\alias{p0}
\alias{caller}
\alias{callers}
\alias{failsafe}
\title{Re-exports from Package \code{uj}}
\usage{
g(g, ...)
g0(...)
p0(...)

N()
av(...)
f0(x, y, n)

caller(.scl = TRUE)
callers(.vec = TRUE)

failsafe(x, .def = "uj.failsafe.err")

re_exports()
}
\description{
Re-exports from Package \code{uj}
}
\section{Functions}{
\itemize{
\item \code{g()}: See \code{\link[uj]{g}}.

\item \code{N()}: See \code{\link[uj]{N}}.

\item \code{av()}: See \code{\link[uj]{av}}.

\item \code{f0()}: See \code{\link[uj]{f0}}.

\item \code{g0()}: See \code{\link[uj]{g0}}.

\item \code{p0()}: See \code{\link[uj]{p0}}.

\item \code{caller()}: See \code{\link[uj]{caller}}.

\item \code{callers()}: See \code{\link[uj]{callers}}.

\item \code{failsafe()}: See \code{\link[uj]{failsafe}}.

}}
\seealso{
Other utils:
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{ssplit}()},
\code{\link{trm}()}
}
\concept{utils}
