% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\encoding{UTF-8}
\name{trm}
\alias{trm}
\title{Glue and Trim Whitespace from Both Sides}
\usage{
trm(..., d = " ")
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=g]{g}}
  \describe{
    \item{\code{}}{}
  }}
}
\value{
A character scalar.
}
\description{
Glue and Trim Whitespace from Both Sides
}
\examples{
.trm(" Just" ,  "a" ,  "little" ,  "thing ")
.trm(" Just ", " a ", " little ", " thing ")
.trm(" ", letters, LETTERS, 0:9, " ", d = "")
}
\seealso{
Other utils: 
\code{\link{choose_from}()},
\code{\link{fmt_errs}()},
\code{\link{glu}()},
\code{\link{match_bg}()},
\code{\link{ok_fmt}()},
\code{\link{re_exports}()},
\code{\link{ssplit}()}
}
\concept{utils}
